Move internal/cluster to pkg/cluster
This avoids having two cluster packages (in internal and pkg). Since
pkg/cluster is used by other public functions, we need to keep it
public.
