+++ pkg/cluster/id_test.go
@@ -23,7 +23,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/submariner-io/subctl/internal/cluster"
+	"github.com/submariner-io/subctl/pkg/cluster"
 )
 
 var _ = Describe("TestClusterIDs", func() {
